Add simple way to toggle cells
When we find any marker entities with the katgolEnable or katgolDisable
tags, we look for the closest cell (marker with katgolCell tag) and set
its isAlive value to the appropriate value.

Summoning these control markers should be the primary way to manipulate
the state of the game. Currently the markers are only summoned by the
debug spawn eggs.